back/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dj-rest-auth==6.0.0
Django==4.2.9
django-allauth==0.63.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
finance-datareader==0.9.90
fonttools==4.51.0
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
numpy==1.26.4
OpenDartReader==0.2.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tenacity==8.3.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
